app/requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.11.1
google-api-python-client
Flask-login
Werkzeug==0.9.6
Flask-SQLAlchemy==2.1
PyMySQL==0.7.9
psycopg2==2.6.1